ws: Split out generic session functionality
Move all utilities, PAM session opening, and the cockpit protocol
with ws into session-utils.[hc], i. e. everything which is not specific
to the authentication type.

This makes it much easier to add new authentication commands.
