+++ src/ws/Makefile-ws.am
@@ -126,6 +126,8 @@ cockpit_session_SOURCES = \
 	src/common/cockpitframe.h \
 	src/common/cockpitmemory.c \
 	src/common/cockpitmemory.h \
+	src/ws/session-utils.c \
+	src/ws/session-utils.h \
 	src/ws/session.c \
 	$(NULL)
 cockpit_session_LDADD = $(COCKPIT_SESSION_LIBS)
